requirements.txt
django
djangorestframework
markdown
django-filter
chardet
djangorestframework-api-key
cryptography